bug #17195 - allow both text and integer orderseq values to be returned to the pool, with text that cannot be coerced to ints being ignored for now. this is a revision of the fix for issue #17119.

+++ xtupleserver/trunk/dbscripts/triggers/cashrcpt.sql
@@ -32,7 +32,7 @@ BEGIN
     _currId = COALESCE(NEW.cashrcpt_curr_id, basecurrid());
 
      --- clear the number from the issue cache
-    PERFORM clearNumberIssue('CashRcptNumber', NEW.cashrcpt_number::INTEGER);
+    PERFORM clearNumberIssue('CashRcptNumber', NEW.cashrcpt_number);
   ELSE
     _currId = NEW.cashrcpt_curr_id;
   END IF;
